requirements.txt
pyautogui
matplotlib
pillow
google-genai
python-dotenv